refactor: move axes_utils to types/xy_chart/utils
