+++ src/lib/series/domains/x_domain.ts
@@ -1,4 +1,4 @@
-import { isCompleteBound, isLowerBound, isUpperBound } from '../../axes/axis_utils';
+import { isCompleteBound, isLowerBound, isUpperBound } from '../../../types/xy_chart/utils/axis_utils';
 import { compareByValueAsc, identity } from '../../../utils/commons';
 import { computeContinuousDataDomain, computeOrdinalDataDomain, Domain } from '../../../utils/domain';
 import { ScaleType } from '../../../utils/scales/scales';
