+++ src/lib/series/domains/y_domain.ts
@@ -1,7 +1,7 @@
 import { BasicSeriesSpec, DomainRange } from '../specs';
 import { GroupId, SpecId } from '../../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../../../utils/scales/scales';
-import { isCompleteBound, isLowerBound, isUpperBound } from '../../axes/axis_utils';
+import { isCompleteBound, isLowerBound, isUpperBound } from '../../../types/xy_chart/utils/axis_utils';
 import { BaseDomain } from './domain';
 import { RawDataSeries } from '../series';
 import { computeContinuousDataDomain } from '../../../utils/domain';
